created initial boot file and added empty blackjack/bust boolean methods
