+++ ruby/blackjack/tests.rb
@@ -1,0 +1,38 @@
+require 'test/unit'
+
+class CardTest < Test::Unit::TestCase
+  def setup
+    @card = Card.new(:hearts, :ten, 10)
+  end
+  
+  def test_card_suite_is_correct
+    assert_equal @card.suite, :hearts
+  end
+
+  def test_card_name_is_correct
+    assert_equal @card.name, :ten
+  end
+  def test_card_value_is_correct
+    assert_equal @card.value, 10
+  end
+end
+
+class DeckTest < Test::Unit::TestCase
+  def setup
+    @deck = Deck.new
+  end
+  
+  def test_new_deck_has_52_playable_cards
+    assert_equal @deck.playable_cards.size, 52
+  end
+  
+  def test_dealt_card_should_not_be_included_in_playable_cards
+    card = @deck.deal_card
+    assert(@deck.playable_cards.include?(card))
+  end
+
+  def test_shuffled_deck_has_52_playable_cards
+    @deck.shuffle
+    assert_equal @deck.playable_cards.size, 52
+  end
+end
